requirements.txt
PyYAML
flask
pytest